refactor(calendar): 트랜잭션 경계를 Writer로 이동하고 Service tx 제거
CalendarEventWriter가 CalendarEventReader를 주입해 update/delete를
eventId 기반 load+소유검증+변경으로 캡슐화(자기 @Transactional).
CalendarEventService의 클래스 @Transactional과 Reader 의존을 제거.
